+++ src/Resources/contao/languages/en/modules.php
@@ -20,3 +20,7 @@ $GLOBALS['TL_LANG']['MOD']['workflow_history'][1] = 'Workflow state history';
 $GLOBALS['TL_LANG']['MOD']['tl_workflow_step']       = 'Step';
 $GLOBALS['TL_LANG']['MOD']['tl_workflow_transition'] = 'Transition';
 $GLOBALS['TL_LANG']['MOD']['tl_workflow_action']     = 'Action';
+
+$GLOBALS['TL_LANG']['FMD']['workflow']               = 'Workflow';
+$GLOBALS['TL_LANG']['FMD']['workflow_transition'][0] = 'Workflow transition';
+$GLOBALS['TL_LANG']['FMD']['workflow_transition'][1] = 'Module processes a workflow transition for an entity.';